% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jst-package.R
\docType{package}
\name{jst-package}
\alias{jst}
\alias{jst-package}
\title{jst: JSON Value Types via S7 Classes}
\description{
Provides S7 classes for representing JSON value types (null, boolean, number, string, array, object). Typed JSON values can be constructed with json_null(), json_boolean(), json_number(), json_string(), json_array(), and json_object(), and serialised to JSON text via format() / print().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pacbtz/jst}
  \item Report bugs at \url{https://github.com/pacbtz/jst/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
